pyproject.toml
[project]
name = "ceartin-mp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.5.0",
    "evaluate[pytorch,torch]>=0.4.6",
    "google-genai>=1.65.0",
    "panphon>=0.22.2",
    "tglstemmer",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers[torch]>=5.2.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=8.0.0",
    "pre-commit>=4.5.1",
    "python-dotenv>=1.2.1",
]

[tool.uv.sources]
tglstemmer = { git = "https://github.com/andrianllmm/tagalog-stemmer.git", rev = "main" }
